requirements.txt
# PPC Suite Requirements

# Web Framework
streamlit>=1.28.0,<2.0.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0

# Machine Learning (for clustering)
scikit-learn>=1.3.0,<2.0.0

# HTTP Requests (for API calls)
requests>=2.31.0

# Excel Support
xlsxwriter>=3.1.0
openpyxl>=3.1.0

# Optional: Firebase (if using Firebase features)
# firebase-admin>=6.2.0
# Database Drivers
psycopg2-binary==2.9.9
python-dotenv==1.0.1
supabase>=2.0.0
plotly>=5.17.0,<6.0.0

# Authentication
bcrypt>=4.0.1

# PDF Generation
fpdf2>=2.7.0
html2image>=2.0.0
kaleido>=0.2.0

# SP-API Pipeline
requests-aws4auth==1.3.1
apscheduler==3.10.4
pytest==8.3.0
pytest-mock==3.14.0
